requirements.txt
fastapi[standard]==0.115.2
uvicorn==0.32.0
langchain-text-splitters==0.3.0
langchain-community==0.3.2
langchain-openai==0.2.2
python-dotenv==1.0.1
langchain-core==0.3.12
httpx==0.27.2
langgraph==0.2.38
colorama==0.4.6
